docker compose #112
途中なので注意

----
1.docker-compose up -d でエラー

WARNING: The http_proxy variable is not set. Defaulting to a blank string.
WARNING: The https_proxy variable is not set. Defaulting to a blank string.
ERROR: The Compose file './docker-compose.yml' is invalid because:
services.db.volumes contains an invalid type, it should be an array
services.emaki.volumes contains an invalid type, it should be an array
services.emaki.ports contains an invalid type, it should be an array
services.db.ports contains an invalid type, it should be an array
services.emaki.depends_on contains an invalid type, it should be an array
services.emaki.build contains unsupported option: 'image'

----
2.compass compile をDockerfileに埋め込んだ

が、実装中に叩くとき不便なのでbin/配下に別途作る